Embedded/Websocket/requirements.txt
pyyaml
websockets
python-dotenv